Refactor FuzzerFactory into ProcessorTool
Change class name in a first step of unifying the tool classes.
Refactor all fuzzer generation logic and move it in the tool class.
Adjust visibility of class members.
